Support underpromotion on Sentio e-boards
Sequence to underpromote:
1. pick up pawn to promote
2. pick up king of same color
3. put down king --> promotion piece changes to knight
4. put down promotion piece on promotion square
Do steps 2+3 twice to change the promotion piece to a bishop.
Do steps 2+3 three times to change the promotion piece to a rook.
Do steps 2+3 four times to change the promotion piece back to a queen.
The promotion piece change is only valid for the next promotion, any other move resets the promotion piece back to a queen.